requirements.txt
altgraph==0.17.4
chardet==5.2.0
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
openpyxl==3.1.5
packaging==24.2
pefile==2023.2.7
pillow>=12.0.0
pyinstaller>=6.10.0
pyinstaller-hooks-contrib==2024.3
PyMuPDF==1.26.7
pywin32-ctypes==0.2.2
reportlab==4.1.0
